ref: Moved all instance custom functions to its own file
