+++ Environment/Libraries/Globals.cpp
@@ -26,28 +26,6 @@ namespace RbxStu {
         return 1;
     }
 
-    int gethui(lua_State *L) {
-        // Equivalent to cloneref(cloneref(cloneref(game):GetService("CoreGui")).RobloxGui)
-        // Excessive clonereffing, I made the cloneref, i will use all of it!
-        // - Dottik
-        lua_getglobal(L, "cloneref");
-        lua_getglobal(L, "game");
-        lua_call(L, 1, 1);
-        lua_getfield(L, -1, "GetService");
-        lua_pushvalue(L, -2);
-        lua_pushstring(L, "CoreGui");
-        lua_call(L, 2, 1);
-        lua_remove(L, 1); // CoreGui is alone on the stack.
-        lua_getglobal(L, "cloneref");
-        lua_pushvalue(L, 1);
-        lua_call(L, 1, 1);
-        lua_getfield(L, -1, "RobloxGui");
-        lua_getglobal(L, "cloneref");
-        lua_pushvalue(L, -2);
-        lua_call(L, 1, 1);
-        return 1;
-    }
-
     int httpget(lua_State *L) {
         const std::string url = luaL_checkstring(L, 1);
 
@@ -82,14 +60,6 @@ namespace RbxStu {
         return lua_yield(L, 1);
     }
 
-    int fireproximityprompt(lua_State *L) {
-        Utilities::checkInstance(L, 1, "ProximityPrompt");
-
-        const auto proximityPrompt = *static_cast<std::uintptr_t **>(lua_touserdata(L, 1));
-        reinterpret_cast<RbxStu::StudioFunctionDefinitions::r_RBX_ProximityPrompt_onTriggered>(
-                RobloxManager::GetSingleton()->GetRobloxFunction("RBX::ProximityPrompt::onTriggered"))(proximityPrompt);
-        return 0;
-    }
 
     int require(lua_State *L) {
         Utilities::checkInstance(L, 1, "ModuleScript");
@@ -170,9 +140,6 @@ luaL_Reg *Globals::GetLibraryFunctions() {
     // WARNING: you MUST add nullptr at the end of luaL_Reg declarations, else, Luau will choke.
     auto *reg = new luaL_Reg[]{{"isluau", RbxStu::isluau},
                                {"httpget", RbxStu::httpget},
-                               {"gethui", RbxStu::gethui},
-                               {"fireproximityprompt", RbxStu::fireproximityprompt},
-
                                {"require", RbxStu::require},
 
                                {"isrbxactive", RbxStu::isrbxactive},
