+++ Environment/Libraries/Instance.hpp
@@ -1,0 +1,12 @@
+//
+// Created by Pixeluted on 22/08/2024.
+//
+
+#pragma once
+#include "Environment/EnvironmentManager.hpp"
+
+class Instance final : public Library {
+public:
+    std::string GetLibraryName() override;
+    luaL_Reg *GetLibraryFunctions() override;
+};
